Refactor related functions to a separate companion object to have access to own logger. Rename tests.
