+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/persistence/CordaPersistence.kt
@@ -80,12 +80,11 @@ class CordaPersistence(
             error("Was not expecting to find existing database transaction on current strand when setting database: ${Strand.currentStrand()}, $it")
         }
         _contextDatabase.set(this)
-        // Check not in read-only mode.
         transaction {
             check(!connection.metaData.isReadOnly) { "Database should not be readonly." }
-            validateOrFixIncorrectInfoHostsTableColumnName(connection)
-            checkCorrectAttachmentsContractsTableName(connection)
-            checkCorrectCheckpointTypeOnPostgres(connection)
+            DatabaseMigration.validateOrAlterAttachmentsContractsTableName(connection)
+            DatabaseMigration.validateOrAlterInfoHostsTableColumnName(connection)
+            DatabaseMigration.checkCorrectCheckpointTypeOnPostgres(connection)
         }
     }
 
@@ -274,72 +273,3 @@ private fun Throwable.hasSQLExceptionCause(): Boolean =
 
 class CouldNotCreateDataSourceException(override val message: String?, override val cause: Throwable? = null) : Exception()
 
-class DatabaseIncompatibleException(override val message: String?, override val cause: Throwable? = null) : Exception()
-
-private fun checkCorrectAttachmentsContractsTableName(connection: Connection) {
-    val correctName = "NODE_ATTACHMENTS_CONTRACTS"
-    val incorrectV30Name = "NODE_ATTACHMENTS_CONTRACT_CLASS_NAME"
-    val incorrectV31Name = "NODE_ATTCHMENTS_CONTRACTS"
-
-    fun warning(incorrectName: String, version: String) = "The database contains the older table name $incorrectName instead of $correctName, see upgrade notes to migrate from Corda database version $version https://docs.corda.net/head/upgrade-notes.html."
-
-    if (!connection.metaData.getTables(null, null, correctName, null).next()) {
-        val metaData = connection.metaData
-        if (connection.metaData.getTables(null, null, incorrectV30Name, null).next()) {
-            if (metaData.databaseProductName == "H2") {
-                connection.createStatement().execute("ALTER TABLE $incorrectV30Name RENAME TO $correctName")
-                connection.commit()
-            } else {
-                throw DatabaseIncompatibleException(warning(incorrectV30Name, "3.0"))
-            }
-        }
-        if (connection.metaData.getTables(null, null, incorrectV31Name, null).next()) {
-            if (metaData.databaseProductName == "H2") {
-                connection.createStatement().execute("ALTER TABLE $incorrectV31Name RENAME TO $correctName")
-                connection.commit()
-            } else {
-                throw DatabaseIncompatibleException(warning(incorrectV31Name, "3.1"))
-            }
-        }
-    }
-}
-
-private fun validateOrFixIncorrectInfoHostsTableColumnName(connection: Connection) = validateOrRenameColumn(connection, "NODE_INFO_HOSTS",  "HOST_NAME", "HOST")
-
-private fun validateOrRenameColumn(connection: Connection, table: String, correctColumnName: String, incorrectColumnName: String) {
-
-    fun warning() = "The database table $table contains the older column name $incorrectColumnName instead of $correctColumnName, " +
-            "see upgrade notes to migrate from Corda database version https://docs.corda.net/head/upgrade-notes.html."
-    val metaData = connection.metaData
-    val resultSet = metaData.getColumns(null, null, table, incorrectColumnName)
-    if (resultSet.next()) {
-        val name = resultSet.getString("COLUMN_NAME")
-        if (name == incorrectColumnName) {
-            if (metaData.databaseProductName == "H2") {
-                try {
-                    connection.createStatement().execute("ALTER TABLE $table ALTER COLUMN $incorrectColumnName RENAME TO $correctColumnName")
-                    connection.commit()
-                } catch (exp: SQLException) {
-                    throw DatabaseIncompatibleException("Failed to upgrade table $table. " + warning())
-                }
-            } else {
-                throw DatabaseIncompatibleException(warning())
-            }
-        }
-    }
-}
-
-private fun checkCorrectCheckpointTypeOnPostgres(connection: Connection) {
-    val metaData = connection.metaData
-    if (metaData.databaseProductName != "PostgreSQL") {
-        return
-    }
-
-    val result = metaData.getColumns(null, null, "node_checkpoints", "checkpoint_value")
-    if (result.next()) {
-        val type = result.getString("TYPE_NAME")
-        if (type != "bytea") {
-            throw DatabaseIncompatibleException("The type of the 'checkpoint_value' table must be 'bytea', but 'oid' was found. See upgrade notes to migrate from Corda database version 3.1 https://docs.corda.net/head/upgrade-notes.html.")
-        }
-    }
-}
